--desarrollo funcionalidades de interfaz

+++ src/com/taxicop/data/Complaint.java
@@ -1,10 +1,10 @@
 package com.taxicop.data;
 
 public class Complaint {
-	public int RANKING;
+	public float RANKING;
 	public String PLACA;
 	public String DESCRIPCION;
-	public 	Complaint(int r, String p, String d){
+	public 	Complaint(float r, String p, String d){
 		RANKING=r;
 		PLACA=p;
 		DESCRIPCION=d;
